+++ src/ts/store.ts
@@ -30,7 +30,7 @@ namespace CdvPurchase {
     /**
      * Current release number of the plugin.
      */
-    export const PLUGIN_VERSION = '13.8.1';
+    export const PLUGIN_VERSION = '13.8.2';
 
     /**
      * Entry class of the plugin.